Improve organization and testing
